requirements copy.txt
binaryornot~=0.4.4
bs4~=0.0.2
configobj~=5.0.8
humanize~=4.10.0
jsons~=1.6.3
langchain-community~=0.2.16
opensearch-py~=2.6.0
Pympler~=1.1
pypdf~=4.3.1
python-dotenv~=1.0.1
requests_oauth2client~=1.5.3
requests-oauthlib~=2.0.0
sentence-transformers~=3.3.1
shiny~=1.2.1
git+https://github.com/posit-dev/py-shinyswatch.git@256d763
isi-util==0.2a2
shabti-util==0.2a2
python-keycloak~=5.1.1
httpx~=0.23.2